+++ admin/metaboxes/link_url/metabox-link_url.php
@@ -46,7 +46,6 @@ function pw_link_url_meta_ui(){
 ////////////// SAVE POST //////////////
 function pw_link_url_meta_save($post_id){
 
-	pw_log( 'pw_link_url_meta_save: '. $post_id );
 	// Stop autosave to preserve meta data
 	if ( wp_is_post_autosave( $post_id ) || wp_is_post_revision( $post_id ) )
         return $post_id;