+++ CHANGELOG.md
@@ -1,4 +1,18 @@
-# Postworld Changelog
+# Postworld // Changelog
+
+### Version 1.7.7
+- Refactor Icon Core
+    + Require `iconset` in modules
+    + Add `iconset` key registry in postworld config
+```php
+'iconsets'  =>  array(
+    'required'  =>  array(
+        'icomoon',
+        'glyphicon-halflings',
+        ),
+    ),
+```
+
 
 ### Version 1.7.5
 - Restructured method of using `postworldAdmin` AngularJS module to Bootstrap metaboxes in WordPress Admin